backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-multipart==0.0.12

google-cloud-bigquery==3.27.0
google-cloud-bigquery-migration==0.11.1
google-cloud-storage==2.18.2

# Vertex AI
google-cloud-aiplatform==1.79.0